+++ cli/src/commands/unlink.rs
@@ -1,0 +1,19 @@
+use anyhow::Result;
+use clap::Parser;
+
+use crate::util::{global_config, paths};
+
+#[derive(Parser)]
+pub struct Opts {}
+
+impl Opts {
+	pub async fn execute(&self, _ctx: &cli_core::Ctx) -> Result<()> {
+		let project_root = paths::project_root()?;
+		global_config::try_mutate_global(|config| {
+			config.project_roots.remove(&project_root);
+			Ok(())
+		})
+		.await?;
+		Ok(())
+	}
+}